+++ CHANGES.rst
@@ -10,9 +10,13 @@ Changelog
   As an alternative, you can specify the ``protocol`` in the configuration
   file.
 
-* Change the return value of ``Model.browse()`` method with an empty list.
-  It returns an empty ``RecordList`` except if some other argument is
-  provided (e.g. ``all_users = model('res.users').browse([], limit=None)``).
+* Change the return value of ``Model.browse()`` method if search domain is
+  an empty list.  It returns an empty ``RecordList`` except if some other
+  argument is provided (e.g.
+  ``all_users = model('res.users').browse([], limit=None)``).
+
+* Change the return value of ``Client.read()`` and ``Model.read()`` methods
+  if search domain is an empty list:  it returns ``False``.
 
 * Improve error formatting for recent Odoo versions, in interactive mode.